requirements.txt
python-vlc
pynput
scipy
numpy
